Share methods between losses_by_asset and dmg_by_asset
